refactor: move util.ts and util_string.ts to utils folder
