+++ src/ts/drive/backuplocal.ts
@@ -6,7 +6,7 @@ import { isTauri } from "src/ts/platform"
 import { decodeRisuSave, encodeRisuSaveLegacy } from "../storage/risuSave";
 import { getDatabase, setDatabaseLite } from "../storage/database.svelte";
 import { relaunch } from "@tauri-apps/plugin-process";
-import { sleep } from "../util";
+import { sleep } from "../utils/util";
 import { hubURL } from "../characterCards.svelte";
 
 export async function SaveLocalBackup(){
