+++ src/ts/gui/colorscheme.ts
@@ -1,7 +1,7 @@
 import { get, writable } from "svelte/store";
 import { getDatabase, setDatabase } from "../storage/database.svelte";
 import { downloadFile } from "../globalApi.svelte";
-import { bufferToText, openFilePicker } from "../util";
+import { bufferToText, openFilePicker } from "../utils/util";
 import { alertError } from "../alert.svelte";
 import { isLite } from "../lite";
 import { CustomCSSStore, SafeModeStore } from "../stores.svelte";
