+++ src/ts/process/index.svelte.ts
@@ -9,7 +9,7 @@ import { ChatTokenizer, tokenize, tokenizeNum } from "../tokenizer";
 import { language } from "../../lang";
 import { alertError } from "../alert.svelte";
 import { loadLoreBookV3Prompt } from "./lorebook.svelte";
-import { isLastCharPunctuation, trimUntilPunctuation, parseToggleSyntax } from "../util";
+import { isLastCharPunctuation, trimUntilPunctuation, parseToggleSyntax } from "../utils/util";
 import { getAuthorNoteDefaultText } from "./prompt";
 import { findCharacterbyId } from "../characters.svelte";
 import { getPersonaPrompt } from "../persona";
