+++ src/ts/process/models/nai.ts
@@ -2,7 +2,7 @@ import { getDatabase, setDatabase } from "src/ts/storage/database.svelte"
 import type { OpenAIChat } from "../index.svelte"
 import { globalFetch } from "src/ts/fetch"
 import { alertError, alertInput, alertNormal, alertWait } from "src/ts/alert.svelte"
-import { sleep } from "src/ts/util"
+import { sleep } from "src/ts/utils/util"
 import { getUserName } from "src/ts/persona"
 
 export function stringlizeNAIChat(formated:OpenAIChat[], char:string, continued: boolean){
