+++ src/ts/process/request/google.ts
@@ -1,8 +1,8 @@
-import { textifyReadableStream } from "src/ts/util"
+import { textifyReadableStream } from "src/ts/utils/util"
 import { globalFetch, fetchNative } from "src/ts/fetch"
 import { LLMFlags, LLMFormat } from "src/ts/model/modellist"
 import { getDatabase, setDatabase } from "src/ts/storage/database.svelte"
-import { simplifySchema } from "src/ts/util"
+import { simplifySchema } from "src/ts/utils/util"
 import { v4 } from "uuid"
 import { setInlayAsset, writeInlayImage } from "../files/inlays"
 import { extractJSON, getGeneralJSONSchema } from "../templates/jsonSchema"
