+++ src/ts/process/request/openAI.ts
@@ -4,14 +4,14 @@ import { getDatabase } from "src/ts/storage/database.svelte"
 import { LLMFlags, LLMFormat } from "src/ts/model/modellist"
 import { strongBan, tokenizeNum } from "src/ts/tokenizer"
 import { getFreeOpenRouterModel } from "src/ts/model/openrouter"
-import { textifyReadableStream } from "src/ts/util"
+import { textifyReadableStream } from "src/ts/utils/util"
 import { globalFetch, fetchNative, addFetchLog } from "src/ts/fetch"
 import { isTauri, isNodeServer, isCapacitor } from "src/ts/platform"
 import type { OpenAIChatFull } from "../index.svelte"
 import { extractJSON, getOpenAIJSONSchema } from "../templates/jsonSchema"
 import { applyChatTemplate } from "../templates/chatTemplate"
 import { supportsInlayImage } from "../files/inlays"
-import { simplifySchema } from "src/ts/util"
+import { simplifySchema } from "src/ts/utils/util"
 import { callTool, decodeToolCall, encodeToolCall } from "../mcp/mcp"
 import { alertError } from "src/ts/alert.svelte";
