+++ src/ts/process/scriptings.ts
@@ -1,4 +1,4 @@
-import { asBuffer } from 'src/ts/util';
+import { asBuffer } from 'src/ts/utils/util';
 import { getChatVar, hasher, setChatVar, getGlobalChatVar, type simpleCharacterArgument, risuChatParser } from "../parser.svelte";
 import { LuaEngine, LuaFactory } from "wasmoon";
 import { getCurrentCharacter, getCurrentChat, getDatabase, setDatabase } from "../storage/database.svelte";
