+++ src/ts/realm.ts
@@ -3,7 +3,7 @@ import { VirtualWriter } from "./globalApi.svelte";
 import { getCurrentCharacter, getDatabase } from "./storage/database.svelte";
 import { type character } from './storage/types/character';
 import { alertClear } from "./alert.svelte";
-import { asBuffer } from "./util";
+import { asBuffer } from "./utils/util";
 
 let pong = false;
