+++ src/ts/utils/util_string.ts
